Add SessionsController#create
A session will be created when a visitor successfully logs in
The user information comes from the omniauth gem:

  def create
    auth_hash = request.env['omniauth.auth']
  ....

The information stored in the session does NOT come from the auth_hash!

  session[:user_id] = user.id
  # .id refers to the auto-incrementing primary key from the Users table

Logging in successfully will redirect to the dashboard index.

TODO: Add a sad path to handle situations where a user clicks the button
and then the OAuth process fails.

Clicking the "Log in with Google button will"

Co-authored-by: Kathy Bui <61847658+Kathybui732@users.noreply.github.com>