Add feature: A user can log in with Google by clicking a button
The user will be created in the database if they have not been already

When a user is logged in, there is a message at the top of every page

The message will not appear if the user is not logged in.
The absence of the message was tested by deleting the cookies in the
local development server and refreshing the page.

TODO: add logout functionality

Co-authored-by: Kathy Bui <61847658+Kathybui732@users.noreply.github.com>